Assetreplacement.py
# Asset Replacement for #5 

This code is a script that defines a class called "Vehicle" that is used to store information about a vehicle, including its name, purchase year, and replacement year. The class has several methods, such as name_check, safety_risk, service_risk, condition_assessment, and manager_adjustment, which are used to update the replacement year based on various inputs.

The script also has a function called validate_input that is used to validate user input, and another function called write_to_csv that is used to write the information about the vehicle to a csv file.

The script starts by prompting the user for the name of a vehicle and then creates an instance of the Vehicle class with that name, along with None for the purchase year and replacement years.

It then calls the name_check method on the vehicle object to see if the vehicle already exists in the csv file. If it does, the script updates the purchase year and replacement years from the csv file. If the vehicle does not exist in the csv file, the script prompts the user for the purchase year and replacement years.

The script then prompts the user for a condition assessment of the vehicle, and calls the condition_assessment method on the vehicle object with the user's input.

Then it prompt the user if a manager adjustment is necessary, if yes then it prompts the user for the manager adjustment and calls the manager_adjustment method on the vehicle object with the user's input.

Finally, the script calls the write_to_csv method on the vehicle object to write the updated information to the csv file and print the final replacement year.